Update to 4.9.7 kernel, the latest LTS release
This update addresses CVE-2016-8405:

"An information disclosure vulnerability in kernel components including the
ION subsystem, Binder, USB driver and networking subsystem could enable a
local malicious application to access data outside of its permission levels.
This issue is rated as Moderate because it first requires compromising a
privileged process".

It should be noted this CVE is applicable largely only to Android.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>